Move function evaluators to separate folder
