+++ src/Altinn.App.Core/Internal/Expressions/ExpressionEvaluator.cs
@@ -3,6 +3,7 @@ using System.Globalization;
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.RegularExpressions;
+using Altinn.App.Core.Internal.Expressions.FunctionEvaluators;
 using Altinn.App.Core.Models;
 using Altinn.App.Core.Models.Expressions;
 using Altinn.App.Core.Models.Layout;
